Level-transition/feature: Had created a new module for creating player environment; and a new module for creating a level
