+++ src/createPlayerEnvironment.js
@@ -1,0 +1,25 @@
+import {createCosmoFactory} from './Entities/Cosmo';
+import PlayerController from './Traits/PlayerController';
+import setupKeyboard from './input';
+import {setupMouseControl} from './debug';
+import createDashboardLayer from './layers/dashboard';
+
+export default function(sprites) {
+    const createCosmo = createCosmoFactory(sprites);
+    const cosmo = createCosmo();
+
+    createPlayerEnv(cosmo);
+    const dashboard = createDashboardLayer(cosmo);
+
+    const input = setupKeyboard(cosmo);
+    setupMouseControl(canvas, cosmo, sprites.camera);
+    input.listenTo(window);
+
+    return dashboard;
+}
+
+function createPlayerEnv (playerEntity) {
+    const playerControl = new PlayerController(playerEntity);
+    playerControl.checkPoint.set(42, 42);
+    playerEntity.addTrait(playerControl);
+}
